requirements.txt
aiohttp==3.9.5
asgiref==3.8.1
datasets==2.20.0
Django==5.0.6
google-generativeai==0.7.1
gTTS==2.5.1
ipython==8.26.0
langchain==0.2.6
langchain-core==0.2.11
langchain-google-genai==1.0.7
langchain-groq==0.1.6
langchain-text-splitters==0.2.2
pillow==10.4.0
gunicorn
